+++ src/Util.hh
@@ -46,7 +46,7 @@ OF THE POSSIBILITY OF SUCH DAMAGE.
 namespace Util
 {
 
-double diffBetweenGroups(const PsiData &psi);
+float diffBetweenGroups(const PsiData &psi);
 void psiToPhi(PhiData &phi, const PsiData &psi);
 void phiToPsi(const PhiData &phi, PsiData &psi);
 void calcTotalSource(const PsiData &source, const PhiData &phi,
